refactor: minor clean-up
