+++ data-protocols/ids/ids-api-multipart-dispatcher-v1/src/main/java/org/eclipse/dataspaceconnector/ids/api/multipart/dispatcher/sender/IdsMultipartSender.java
@@ -58,8 +58,8 @@ import static java.util.concurrent.CompletableFuture.failedFuture;
 /**
  * Abstract class for sending IDS multipart messages.
  *
- * @param <M> the RemoteMessage type sent by the sub class.
- * @param <R> the response type returned by the sub class.
+ * @param <M> the RemoteMessage type sent by the subclass.
+ * @param <R> the response type returned by the subclass.
  */
 abstract class IdsMultipartSender<M extends RemoteMessage, R> implements IdsMessageSender<M, R> {
     private static final String TOKEN_SCOPE = "idsc:IDS_CONNECTOR_ATTRIBUTES_ALL";
